1.8.4 done

+++ src/main/java/com/opstty/job/TallestTree.java
@@ -5,7 +5,7 @@ import com.opstty.mapper.TallestMapper;
 import com.opstty.reducer.TallestReducer;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.FloatWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -26,7 +26,7 @@ public class TallestTree {
         job.setCombinerClass(TallestReducer.class);
         job.setReducerClass(TallestReducer.class);
         job.setOutputKeyClass(Text.class);
-        job.setOutputValueClass(FloatWritable.class);
+        job.setOutputValueClass(DoubleWritable.class);
         for (int i = 0; i < otherArgs.length - 1; ++i) {
             FileInputFormat.addInputPath(job, new Path(otherArgs[i]));
         }
